Move module to webgl directory
