+++ config/tsconfig-strict.json
@@ -278,8 +278,8 @@
     "../src/ol/VectorRenderTile.js",
     "../src/ol/VectorTile.js",
     "../src/ol/View.js",
-    "../src/ol/WebGLCanvas.js",
     "../src/ol/webgl/BaseTileRepresentation.js",
+    "../src/ol/webgl/Canvas.js",
     "../src/ol/webgl/Helper.js",
     "../src/ol/webgl/PaletteTexture.js",
     "../src/ol/webgl/PostProcessingPass.js",
